fix: merge conflict 해결 - SwaggerConfig 2개 -> 1개로 통합
