+++ backend/src/main/java/org/prgrms/devconnect/domain/config/SwaggerConfig.java
@@ -6,6 +6,7 @@ import io.swagger.v3.oas.models.OpenAPI;
 import io.swagger.v3.oas.models.info.Info;
 import io.swagger.v3.oas.models.security.SecurityRequirement;
 import io.swagger.v3.oas.models.security.SecurityScheme;
+import org.springdoc.core.models.GroupedOpenApi;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
@@ -15,9 +16,11 @@ public class SwaggerConfig {
   @Bean
   public OpenAPI customOpenAPI() {
     return new OpenAPI()
-        .info(new Info().title("DevConnect")
+        .info(new Info()
+            .title("springdoc-openapi")
             .version("1.0")
-            .description("API Description"))
+            .description("DevConnector swagger-ui 화면입니다.")
+        )
         .components(new Components()
             .addSecuritySchemes("bearer-jwt",
                 new SecurityScheme()
@@ -25,5 +28,70 @@ public class SwaggerConfig {
                     .scheme("bearer")
                     .bearerFormat("JWT")))
         .addSecurityItem(new SecurityRequirement().addList("bearer-jwt"));
+  }
+
+  @Bean
+  public GroupedOpenApi getAllAPI() {
+    return GroupedOpenApi.builder()
+        .group("all")
+        .pathsToMatch("/api/v1/**")
+        .build();
+  }
+
+  @Bean
+  public GroupedOpenApi getAlarmAPI() {
+    return GroupedOpenApi.builder()
+        .group("alarms")
+        .pathsToMatch("/api/v1/alarms/**")
+        .build();
+  }
+
+  @Bean
+  public GroupedOpenApi getBoardAPI() {
+    return GroupedOpenApi.builder()
+        .group("boards")
+        .pathsToMatch("/api/v1/boards/**")
+        .build();
+  }
+
+  @Bean
+  public GroupedOpenApi getChattingAPI() {
+    return GroupedOpenApi.builder()
+        .group("chatting")
+        .pathsToMatch("/api/v1/chat/**")
+        .build();
+  }
+
+  @Bean
+  public GroupedOpenApi getInterestAPI() {
+    return GroupedOpenApi.builder()
+        .group("interests")
+        .pathsToMatch("/api/v1/interests/**")
+        .build();
+  }
+
+
+  @Bean
+  public GroupedOpenApi getJobpostAPI() {
+    return GroupedOpenApi.builder()
+        .group("job-posts")
+        .pathsToMatch("/api/v1/job-posts/**")
+        .build();
+  }
+
+  @Bean
+  public GroupedOpenApi getMemberAPI() {
+    return GroupedOpenApi.builder()
+        .group("members")
+        .pathsToMatch("/api/v1/members/**")
+        .build();
+  }
+
+  @Bean
+  public GroupedOpenApi getTechStackAPI() {
+    return GroupedOpenApi.builder()
+        .group("tech-stacks")
+        .pathsToMatch("/api/v1/tech-stacks/**")
+        .build();
   }
 }
